Task1

+++ src/main/java/com/btvn/ss16/model/OrderDetail.java
@@ -33,5 +33,5 @@ public class OrderDetail {
     private Integer quantity;
 
     @Column(nullable = false)
-    private Long unitPrice;
+    private Double unitPrice;
 }
